backend/requirements.txt
fastapi
uvicorn
sqlalchemy
authlib
python-jose
passlib
cryptography
python-multipart
python-dotenv
google-generativeai
requests
markdown-it-py
beautifulsoup4
openpyxl
python-pptx
chardet
faiss-cpu
chromadb
sentence-transformers
httpx
